// src/components/schedular/types.d.ts
interface CalendarEvent {
    event_id: number | string;
    title: string;
    start: Date;
    end: Date;
    disabled?: boolean;
    color?: string;
    editable?: boolean;
    deletable?: boolean;
    draggable?: boolean;
    allDay?: boolean;
}
// ...
export interface EventRendererProps extends Pick<React.HTMLAttributes<HTMLElement>, "draggable" | "onDragStart" | "onDragEnd" | "onDragOver" | "onDragEnter" | "onClick"> {
    event: ProcessedEvent;
    view?: View;
}
// ...
export type ProcessedEvent = CalendarEvent & Record<string, any>;